requirements.txt
absl-py==0.7.1
biopython==1.77
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
grpcio==1.25.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.0.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pysam==0.15.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard-plugin-wit==1.7.0
torch==1.5.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
